move Replicator ifc to repos/db
